mock sciencebeam parser
